+++ api-admin/controllers/notificationlog/index.js
@@ -1,10 +1,12 @@
 
 import express from 'express';
+import create from './create';
 import list from './list';
 
 let router = express.Router();
 
 router.route('/notificationLog')
-    .get(list);
+    .get(list)
+    .post(create);
 
 module.exports = router;